requirements.txt
flask==3.1.0
python-dotenv==1.1.0
requests==2.32.5
certifi==2026.1.4
charset-normalizer==3.4.4
idna==3.11
urllib3==2.6.3
